[CL-1941] Move ContentBuilderEditModePreview/EditModePreview to containers as FullscreenPreview
